+++ src/@next/components/atoms/CartFooter/types.ts
@@ -1,6 +1,0 @@
-export interface IProps {
-  subtotalPrice: React.ReactNode;
-  shippingPrice?: React.ReactNode;
-  discountPrice?: React.ReactNode;
-  totalPrice: React.ReactNode;
-}