+++ src/@next/components/atoms/index.ts
@@ -30,3 +30,5 @@ export * from "./SocialMediaIcon";
 export * from "./StripeInputElement";
 export * from "./Tile";
 export * from "./TileGrid";
+export * from "./CartCostsSummary";
+export * from "./OfflinePlaceholder";