+++ src/@next/components/molecules/CardHeader/stories.tsx
@@ -31,7 +31,7 @@ storiesOf("@components/molecules/CardHeader", module)
     };
     return renderHeader(PROPS);
   })
-  .add("with custom icon", () => {
+  .add("with custom close icon", () => {
     const icon = (
       <IconButton
         testingContext="test"
@@ -40,7 +40,19 @@ storiesOf("@components/molecules/CardHeader", module)
         onClick={action("edit")}
       />
     );
-    const PROPS = { ...DEFAULT_PROPS, customIcon: icon };
+    const PROPS = { ...DEFAULT_PROPS, closeIcon: icon };
+    return renderHeader(PROPS);
+  })
+  .add("with prefix", () => {
+    const icon = (
+      <IconButton
+        testingContext="test"
+        name="arrow_back"
+        size={24}
+        onClick={action("onArrowBack")}
+      />
+    );
+    const PROPS = { ...DEFAULT_PROPS, prfix: icon };
     return renderHeader(PROPS);
   })
   .add("with paragraph text style", () => {
@@ -49,7 +61,7 @@ storiesOf("@components/molecules/CardHeader", module)
     return renderHeader(PROPS);
   })
   .add("with big text size", () => {
-    const customIcon = (
+    const closeIcon = (
       <IconButton
         testingContext="test"
         name="trash"
@@ -58,6 +70,6 @@ storiesOf("@components/molecules/CardHeader", module)
       />
     );
     const titleSize: TitleSize = "lg";
-    const PROPS = { ...DEFAULT_PROPS, customIcon, divider: true, titleSize };
+    const PROPS = { ...DEFAULT_PROPS, closeIcon, divider: true, titleSize };
     return renderHeader(PROPS);
   });