+++ src/@next/components/molecules/CardHeader/test.tsx
@@ -52,7 +52,7 @@ describe("<CardHeader />", () => {
     expect(paragraph.contains(children)).toBe(true);
   });
 
-  it("should render close icon when `onHide` prop is defined and `customIcon` is not defined", () => {
+  it("should render close icon when `onHide` prop is defined and `closeIcon` is not defined", () => {
     const onHide = jest.fn();
     const closeIcon = renderHeader({
       ...DEFAULT_PROPS,
@@ -66,19 +66,25 @@ describe("<CardHeader />", () => {
     expect(onHide).toHaveBeenCalled();
   });
 
-  it("should render custom icon when `customIcon` prop is defined", () => {
-    const customIcon = <IconButton testingContext="test" name="edit" />;
-    const icon = renderHeader({ ...DEFAULT_PROPS, customIcon }).find(
-      IconButton
-    );
+  it("should render custom close icon when `closeIcon` prop is defined", () => {
+    const closeIcon = <IconButton testingContext="test" name="edit" />;
+    const icon = renderHeader({ ...DEFAULT_PROPS, closeIcon }).find(IconButton);
 
     expect(icon.exists()).toEqual(true);
     expect(icon.prop("name")).toEqual("edit");
   });
 
+  it("should render prefix as icon", () => {
+    const prefix = <IconButton testingContext="test" name="arrow_back" />;
+    const icon = renderHeader({ ...DEFAULT_PROPS, prefix }).find(IconButton);
+
+    expect(icon.exists()).toEqual(true);
+    expect(icon.prop("name")).toEqual("arrow_back");
+  });
+
   it("should render border-bottom when `divider` prop is set to true", () => {
     const header = mount(<CardHeader {...DEFAULT_PROPS} divider />).find(
-      S.Header
+      S.Wrapper
     );
 
     expect(header).toHaveStyleRule(